Fix threading perfomance of this allocator
The code is very badly written for muli-threading as it locks
globally on the buddy allocator. Fix it by making the two
structures used to implement the cache have locks per structure
of each allocated to make allocation locked globally to each
cache. This will improve threading perfomance and cache loyality
on modern CPUS.